// sipa-boot-sdk/sipa-boot-sdk-js/src/lib/logan/interface.ts
export enum ResultMsg {
  DB_NOT_SUPPORT = "IndexedDB is not supported",
  NO_LOG = "No log exists",
  REPORT_LOG_SUCCESS = "Report success",
  REPORT_LOG_FAIL = "Report fail",
  EXCEED_TRY_TIMES = "Exceed try times",
  EXCEED_LOG_SIZE_LIMIT = "Exceed log size day limit",
}
// ...
export enum LogEncryptMode {
  PLAIN = 0,
  RSA = 1,
}
// ...
export interface LogSuccessHandler {
  (log: LogConfig): void | Promise<void>
}
// ...
export interface ReportResult {
  [key: string]: {
    msg: ResultMsg
    desc?: string
  }
}

/**
 * Log-saving related configs.
 */
export interface LogConfig {
  logContent: string
  encryptVersion: LogEncryptMode
}